c867 Ch. 6.7.1
Read integers from input and store each integer into a vector until -1 is read. Do not store -1 into the vector. Then, if the last value in the vector is negative, output the negative values in the vector, each on a new line. Otherwise, output all values in the vector that are not negative, each on a new line.

Ex: If the input is -81 36 -74 2 -1, the output is:

36
2